+++ Release-Notes.md
@@ -1,5 +1,11 @@
 # PSRest Release Notes
 
+## v0.0.2
+
+Require `Set-RestEnvironment` before other commands.
+
+Avoid .env imported to process environment.
+
 ## v0.0.1
 
 Commands: